Enhance API template: add `<Method>WithParams` for optional params via structs, instead of arguments
Previously, all parameters were passed through method arguments.

This made it difficult to exclude optional parameters such as
the context ID for `Runtime.Evaluate`. In the described case,
specifying `0` was not enough for it to be considered 'empty'.
It was simply treated as a value.

This change arguably makes usage of this package slightly more
verbose as a struct for every method's param will have to be
initialised. Nevertheless, the effects of this is somewhat
counterbalanced by the fact that we had to already do this before
when specifying every argument for a method call (even for params
we do not intend on using). Taken together, this may actually
simplify the API's usage, and make it more idiomatic (by defining
a structure for every method's properties rather than a generic
map interface).

As there were conflicting method names against different types,
e.g. `DeleteCookie`, it was not enough to just append `Params`
to the method name to generate the struct's name. The domain
is prepended for extra safety / to avoid conflicts.

Unfortunately, this change entails importing `gcdapi` to access
the types when using the API. A solution to this may be to
create 'initialisers' against a service,
e.g. `Page.NewDeleteCookieParams(..)`

Another approach may be to use functional options, but that
route may result in unnecessary bloat in the generated library.
This approach is easier to comprehend.

To maintain backwards compatibility, existing methods (using
arguments) will simply proxy the call to the new `<Method>WithParams`.